refactor: Move /browse/map page to /browse
